[cleanup] checking with sonar

+++ benchmarks/src/main/java/uk/ac/manchester/tornado/benchmarks/ComputeKernels.java
@@ -17,8 +17,6 @@
  */
 package uk.ac.manchester.tornado.benchmarks;
 
-import org.apache.lucene.util.LongBitSet;
-
 import uk.ac.manchester.tornado.api.annotations.Parallel;
 import uk.ac.manchester.tornado.api.collections.math.TornadoMath;
 import uk.ac.manchester.tornado.api.collections.types.Byte3;
@@ -64,18 +62,18 @@ public class ComputeKernels {
                 seed = (seed * 0x5DEECE66DL + 0xBL) & ((1L << 48) - 1);
                 // this generates a number between 0 and 1 (with an awful
                 // entropy)
-                float x = ((float) (seed & 0x0FFFFFFF)) / 268435455f;
+                float x = (seed & 0x0FFFFFFF) / 268435455f;
                 // repeat for y
                 seed = (seed * 0x5DEECE66DL + 0xBL) & ((1L << 48) - 1);
                 seed = (seed * 0x5DEECE66DL + 0xBL) & ((1L << 48) - 1);
-                float y = ((float) (seed & 0x0FFFFFFF)) / 268435455f;
+                float y = (seed & 0x0FFFFFFF) / 268435455f;
                 float dist = TornadoMath.sqrt(x * x + y * y);
                 if (dist <= 1.0f) {
                     sum += 1.0f;
                 }
             }
             sum = sum * 4;
-            result[idx] = sum / (float) iter;
+            result[idx] = sum / iter;
         }
     }
